refactor and implement most of getter

+++ ipld/amt/src/root.rs
@@ -9,7 +9,7 @@ use encoding::{
 
 #[derive(PartialEq, Eq, Debug, Default)]
 pub struct Root<'a> {
-    pub(super) height: u64,
+    pub(super) height: u32,
     pub(super) count: u64,
     pub(super) node: Node<'a>,
 }
